Renamed KnotProxy to Knot.
