+++ core/src/test/java/io/knotx/engine/handler/action/CircuitBreakerActionTest.java
@@ -17,15 +17,23 @@
  */
 package io.knotx.engine.handler.action;
 
+import static io.knotx.engine.api.fragment.FragmentResult.DEFAULT_TRANSITION;
+
+import io.knotx.engine.api.fragment.Action;
 import io.knotx.engine.api.fragment.FragmentResult;
 import io.knotx.engine.handler.action.CircuitBreakerActionFactory.CircuitBreakerAction;
+import io.knotx.fragment.Fragment;
 import io.vertx.circuitbreaker.CircuitBreaker;
 import io.vertx.circuitbreaker.CircuitBreakerOptions;
 import io.vertx.circuitbreaker.impl.CircuitBreakerImpl;
-import io.vertx.core.Future;
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
+import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.VertxExtension;
-
+import io.vertx.junit5.VertxTestContext;
+import java.util.concurrent.TimeUnit;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -34,23 +42,20 @@ import org.junit.jupiter.api.extension.ExtendWith;
 @ExtendWith(VertxExtension.class)
 class CircuitBreakerActionTest {
 
+  private static final Fragment FRAGMENT = new Fragment("type", new JsonObject(), "expectedBody");
+  private static final Action DO_ACTION = (fragmentContext, resultHandler) ->
+      new FragmentResult(FRAGMENT, DEFAULT_TRANSITION);
+
+
   @Test
   @DisplayName("Expect operation ends when doAction ends on time.")
-  void expectOperationEnds(Vertx vertx) {
+  void expectOperationEnds(Vertx vertx) throws Throwable {
     // given
-    CircuitBreaker circuitBreaker = new CircuitBreakerImpl("name", vertx,
-        new CircuitBreakerOptions());
-    CircuitBreakerAction tested = new CircuitBreakerAction(circuitBreaker,
-        (fragmentContext, resultHandler) -> resultHandler.handle(
-            Future.succeededFuture(new FragmentResult(null, FragmentResult.DEFAULT_TRANSITION))));
-    // TODO
-    // when
 
-    //then
   }
 
   @Test
-  @DisplayName("Expect fallback transition when doAction timeouts.")
+  @DisplayName("Expect fallback transition when doAction times out.")
   void expectFallback(Vertx vertx) {
     // given
 
@@ -62,6 +67,16 @@ class CircuitBreakerActionTest {
   @Test
   @DisplayName("Expect doAction not applied when circuit is closed.")
   void expectDoActionNotCalled(Vertx vertx) {
+    // given
+
+    // when
+
+    //then
+  }
+
+  @Test
+  @DisplayName("Expect doAction applied when circuit is half-opened.")
+  void expectDoActionAppliedWhenHalfOpen(Vertx vertx) {
     // given
 
     // when
